test(openapi-generator): cover multipart record payloads
